Create person refactor
